+++ src/Grid.tsx
@@ -1,0 +1,7 @@
+import styled from "@emotion/styled";
+
+const Grid = styled.div({
+  display: "grid",
+});
+
+export default Grid;