+++ src/Overlay/Backdrop.tsx
@@ -1,0 +1,9 @@
+import styled from "@emotion/styled";
+
+const Backdrop = styled.div({
+  display: "flex",
+  alignItems: "center",
+  justifyContent: "center",
+});
+
+export default Backdrop;