+++ src/replaceFirstChild.js
@@ -1,8 +1,0 @@
-export default function replaceFirstChild(newChild) {
-  const { firstChild } = this;
-  if (firstChild) {
-    this.replaceChild(newChild, firstChild);
-  } else {
-    this.appendChild(newChild);
-  }
-}