+++ CHANGELOG.md
@@ -1,11 +1,17 @@
+# 1.0.0
+
+- **Breaking:** Rename and update most of optional props
+- Add and export `SuspenseOverlayCore`
+- Enable CSS transitions
+
 # 0.1.2
 
-- Remove demo.gif from the package
+- Remove `demo.gif` from the package
 
 # 0.1.1
 
 - Improve blur. Remove white semi-transparent background.
-- Update README.md
+- Update `README.md`
 
 # 0.1.0